Lista de productos en carrito

+++ lib/models/product.dart
@@ -2,7 +2,7 @@ import 'dart:ffi';
 
 class Product {
   String name;
-  Double price;
+  int price;
   int inStock;
   
   Product({
